hw6

+++ src/services/auth.js
@@ -2,21 +2,23 @@ import { randomBytes } from 'crypto';
 import bcrypt from 'bcrypt';
 import createHttpError from 'http-errors';
 import { UsersCollection } from '../db/models/user.js';
-import { FIFTEEN_MINUTES, THIRTY_DAYS } from '../index.js';
 import { SessionsCollection } from '../db/models/session.js';
 import jwt from 'jsonwebtoken';
 import { sendEmail } from '../utils/sendMail.js';
 import { getEnvVar } from '../utils/getEnvVar.js';
-import { SMTP } from '../constants/index.js';
+import { SMTP,TEMPLATES_DIR, TTL } from '../constants/index.js';
 import handlebars from 'handlebars';
 import path from 'node:path';
 import fs from 'node:fs/promises';
 
 export const registerUser = async (payload) => {
-  const user = await UsersCollection.findOne({ email: payload.email });
+  const user = await UsersCollection.findOne({
+    email: payload.email,
+  });
   if (user) throw createHttpError(409, 'Email in use');
 
   const encryptedPassword = await bcrypt.hash(payload.password, 10);
+
   return await UsersCollection.create({
     ...payload,
     password: encryptedPassword,
@@ -25,61 +27,75 @@ export const registerUser = async (payload) => {
 
 export const loginUser = async (payload) => {
   const user = await UsersCollection.findOne({ email: payload.email });
-  if (!user) {
-    throw createHttpError(401, 'User not found');
-  }
-  const isEqual = await bcrypt.compare(payload.password, user.password);
+  if (!user) throw createHttpError(401, 'Invalid email or password');
 
-  if (!isEqual) {
-    throw createHttpError(401, 'Unauthorized');
-  }
+  const isEqual = await bcrypt.compare(payload.password, user.password);
+  if (!isEqual) throw createHttpError(401, 'Invalid email or password');
+
   await SessionsCollection.deleteOne({ userId: user._id });
 
-  const accessToken = randomBytes(30).toString('base64');
-  const refreshToken = randomBytes(30).toString('base64');
+  const accessToken = randomBytes(50).toString('base64');
+  const refreshToken = randomBytes(50).toString('base64');
 
   return await SessionsCollection.create({
     userId: user._id,
     accessToken,
     refreshToken,
-    accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
-    refreshTokenValidUntil: new Date(Date.now() + THIRTY_DAYS),
+    accessTokenValidUntil: new Date(
+      Date.now() + TTL.ACCESS_TOKEN.FIFTEEN_MINUTES,
+    ),
+    refreshTokenValidUntil: new Date(
+      Date.now() + TTL.REFRESH_TOKEN.THIRTY_DAYS,
+    ),
   });
 };
-export const logoutUser = async (sessionId) => {
-  await SessionsCollection.deleteOne({ _id: sessionId });
+
+const createSession = () => {
+  const accessToken = randomBytes(50).toString('base64');
+  const refreshToken = randomBytes(50).toString('base64');
+
+  return {
+    accessToken,
+    refreshToken,
+    accessTokenValidUntil: new Date(
+      Date.now() + TTL.ACCESS_TOKEN.FIFTEEN_MINUTES,
+    ),
+    refreshTokenValidUntil: new Date(
+      Date.now() + TTL.REFRESH_TOKEN.THIRTY_DAYS,
+    ),
+  };
 };
 
-export const refreshUsersSession = async ({ sessionId, refreshToken }) => {
+export const refreshUserSession = async (sessionId, refreshToken) => {
   const session = await SessionsCollection.findOne({
     _id: sessionId,
     refreshToken,
   });
-
-  if (!session) {
-    throw createHttpError(401, 'Session not found');
-  }
+  if (!session)
+    throw createHttpError(401, 'User not authorized, please log in!');
 
   const isRefreshTokenExpired =
     new Date() > new Date(session.refreshTokenValidUntil);
+  if (isRefreshTokenExpired)
+    throw createHttpError(401, 'A session token has expired.');
 
-  if (isRefreshTokenExpired) {
-    throw createHttpError(401, 'Refresh token expired');
-  }
+  await SessionsCollection.deleteOne({ _id: sessionId, refreshToken });
+  const newSession = createSession();
 
-  const newAccessToken = randomBytes(30).toString('base64');
-
-  session.accessToken = newAccessToken;
-  session.accessTokenValidUntil = new Date(Date.now() + FIFTEEN_MINUTES);
-  await session.save();
-
-  return session;
+  return await SessionsCollection.create({
+    userId: session.userId,
+    ...newSession,
+  });
 };
 
-export const requestResetToken = async (email) => {
+export const logoutUser = async (sessionId) => {
+  await SessionsCollection.deleteOne({ _id: sessionId });
+};
+
+export const sendResetToken = async (email) => {
   const user = await UsersCollection.findOne({ email });
   if (!user) {
-    throw createHttpError(404, 'User not found');
+    throw createHttpError(404, 'User not found!');
   }
 
   const resetToken = jwt.sign(
@@ -89,55 +105,76 @@ export const requestResetToken = async (email) => {
     },
     getEnvVar('JWT_SECRET'),
     {
-      expiresIn: '5m',
+      expiresIn: '10m',
     },
   );
-  const TEMPLATES_DIR = path.resolve('src', 'templates');
+
   const resetPasswordTemplatePath = path.join(
     TEMPLATES_DIR,
-    'reset-password-email.html',
+    'reset-password-email.hbs',
   );
-  const templateSource = (
+
+  const templateMarkup = (
     await fs.readFile(resetPasswordTemplatePath)
   ).toString();
-  const template = handlebars.compile(templateSource);
+
+  const template = handlebars.compile(templateMarkup);
 
   const html = template({
-    name: user.name || 'User',
+    name: user.name,
     link: `${getEnvVar('APP_DOMAIN')}/reset-password?token=${resetToken}`,
   });
+
   try {
     await sendEmail({
       from: getEnvVar(SMTP.SMTP_FROM),
       to: email,
-      subject: 'Reset your password',
+      subject: 'Request to reset your password',
       html,
     });
   } catch (error) {
-    console.error('Email send failed:', error);
     throw createHttpError(
       500,
       'Failed to send the email, please try again later.',
+      { cause: error },
     );
   }
 };
 
-export const resetPassword = async ({ token, password }) => {
-  let payload;
+export const resetPassword = async (token, password) => {
+  let entries;
   try {
-    payload = jwt.verify(token, getEnvVar('JWT_SECRET'));
-  } catch (error) {
-    throw createHttpError(401, 'Token is expired or invalid.');
+    entries = jwt.verify(token, getEnvVar('JWT_SECRET'));
+  } catch (err) {
+    if (err instanceof Error)
+      throw createHttpError(401, 'Token is expired or invalid.', {
+        cause: err,
+      });
+    throw err;
   }
 
-  const user = await UsersCollection.findOne({ email: payload.email });
+  const user = await UsersCollection.findOne({
+    _id: entries.sub,
+    email: entries.email,
+  });
   if (!user) {
-    throw createHttpError(404, 'User not found!');
+    throw createHttpError(404, 'User not found.');
   }
 
-  const hashedPassword = await bcrypt.hash(password, 10);
-  user.password = hashedPassword;
-  await user.save();
+  const isEqual = await bcrypt.compare(password, user.password);
+  if (isEqual) {
+    throw createHttpError(
+      400,
+      'New password cannot be the same as the current password.',
+    );
+  }
 
-  await SessionsCollection.deleteMany({ userId: user._id });
+  const encryptedPassword = await bcrypt.hash(password, 10);
+
+  await UsersCollection.updateOne(
+    { _id: user._id },
+    { password: encryptedPassword },
+  );
+
+  await SessionsCollection.deleteOne({ userId: user._id });
 };
